// src/services/types/api-endpoint-save-request.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export const apiEndpointSaveRequestHttpMethodsEnum = {GET: "GET", POST: "POST", PUT: "PUT", DELETE: "DELETE"} as const;

export type ApiEndpointSaveRequestHttpMethodsEnumKey =
	(typeof apiEndpointSaveRequestHttpMethodsEnum)[keyof typeof apiEndpointSaveRequestHttpMethodsEnum];

export type ApiEndpointSaveRequest = {
	/**
	 * @description Tag name
	 * @minLength 1
	 * @type string
	 */
	tagName: string;
	/**
	 * @description 사용 가능한 HTTP 메서드 목록
	 * @type array
	 */
	httpMethods: ApiEndpointSaveRequestHttpMethodsEnumKey[];
};