// src/services/types/permission-group-response.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export type PermissionGroupResponse = {
	/**
	 * @description 권한 그룹 아이디
	 * @type integer, int64
	 */
	id: number;
	/**
	 * @description 권한 그룹명
	 * @type string
	 */
	name: string;
	/**
	 * @description 설명
	 * @type string | undefined
	 */
	description?: string;
};